requirements.txt
attrs==15.2.0
BeautifulSoup==3.2.1
blinker==1.3
chardet==2.3.0
configglue==1.1.2
configobj==5.0.6
cryptography==1.2.3
debtags==2.0
defer==1.0.6
dirspec==13.10
Django==1.11.11
djangorestframework==3.7.7
dnspython==1.12.0
duplicity==0.7.6
ecdsa==0.13
enum34==1.1.2
feedparser==5.1.3
gyp==0.1
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
nemo-emblems==3.2.3
netifaces==0.10.4
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
paramiko==1.16.0
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
protobuf==2.6.1
psycopg2==2.7.4
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pyinotify==0.9.6
PyJWT==1.3.0
pyOpenSSL==0.15.1
pyparted==3.10.7
pyserial==3.0.1
pysmbc==1.0.15.5
python-apt==1.1.0b1
python-debian==0.1.27
python-xlib==0.14
pytz==2018.3
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
scour==0.32
service-identity==16.0.0
setproctitle==1.1.8
six==1.10.0
Twisted==16.0.0
urllib3==1.13.1
uTidylib==0.2
vboxapi==1.0
virtualenv==15.2.0
wxPython==3.0.2.0
wxPython-common==3.0.2.0
youtube-dl==2016.2.22
zope.interface==4.1.3
